Removed Organization info due to language validation failure
Although we had the best intentions in mind to keep this available, I
was seeing metadata load failures earlier this week due to
System.IdentityModel throwing exceptions on "lang" attributes where the
CultureInfo could not be identified.

Removed the Organization parsing code, as well as the test which checked
for org duplicates.