Add HttpRequestData<T> for notification context
Given that the core library is still targeting .NET Standard, we
can't introduce .NET Core's HttpContext yet. As a bridge to that
point, this solution derives a class with a generic parameter from
the current HttpRequestData, which has a new object property to
hold the context. This is a similar approach to IEnumerator<T> vs.
the object-based IEnumerator. Although in that case, the Current
property is suppressed with the new modifier, whereas this solution
uses a virtual property with a second typed property in the derived
class, so that the instance data can be shared instead of being
duplicated.

The notification handler signatures have been modified to take the
new context object. Implementors will need to cast to their target-
specific type (i.e., HttpContext on Core or HttpRequestBase on
Framework). This can be replaced with a typed parameter if/when this
library becomes Core only.

GetClaims was modified to take a context argument. The reason that
it's not defaulted to null is because that would make it easy to
mistakenly not provide that parameter in the library itself. The
cost is that existing tests now have an extra null argument.